logging init/down now is ref counted severity and facility are now 32bit (up from 16bit)

+++ libs/gkr_log/gkr/log/message.h
@@ -6,15 +6,15 @@ extern "C" {
 
 struct gkr_log_message
 {
-    long long      tid;
-    long long      stamp;
-    unsigned short severity;
-    unsigned short facility;
-    unsigned int   messageLen;
-    const char*    messageText;
-    const char*    threadName;
-    const char*    severityName;
-    const char*    facilityName;
+    long long   tid;
+    long long   stamp;
+    int         severity;
+    int         facility;
+    unsigned    messageLen;
+    const char* messageText;
+    const char* threadName;
+    const char* severityName;
+    const char* facilityName;
 };
 
 #ifdef __cplusplus
